requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==45.0.2
daphne==4.2.0
Django==5.2.1
google-generativeai==0.8.5
hyperlink==21.0.0
incremental==24.7.2
msgpack==1.1.0
pyOpenSSL==25.1.0
redis==6.1.0
Twisted==24.11.0
txaio==23.1.1
youtube-transcript-api==1.0.3
zope.interface==7.2
